services/transcription-service/requirements.txt
fastapi
uvicorn
#openai-whisper
numpy
faster-whisper
webrtcvad-wheels
